# contract/openapi.yaml
openapi: 3.0.0
info:
  title: Kafbat UI Api Service
  contact:
    name: API Support
    email: support@kafbat.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.2.0
...
paths:
...
  /api/clusters/{clusterName}/clientquotas:
    get:
      operationId: listQuotas
      summary: listQuotas
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientQuotas'
      tags:
        - ClientQuotas
    post:
      operationId: upsertClientQuotas
      summary: upsertClientQuotas
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - ClientQuotas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientQuotas'
...
components:
  schemas:
...
    ClientQuotas:
      type: object
      properties:
        user:
          type: string
        clientId:
          type: string
        ip:
          type: string
        quotas:
          type: object
          additionalProperties:
            type: number
            format: decimal
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
        - requestId
      properties:
        code:
          type: integer
          format: int32
          description: Internal error code (can be used for message formatting & localization on UI)
        message:
          type: string
          description: Error message
        timestamp:
          type: number
          format: decimal
          description: Response unix timestamp in ms
        requestId:
          type: string
          description: Unique server-defined request id for convenient debugging
        fieldsErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        stackTrace:
          type: string
    FieldError:
      type: object
      required:
        - fieldName
      properties:
        fieldName:
          type: string
          description: Name of field that violated format
        restrictions:
          type: array
          items:
            type: string
          description: Field format violations description (ex. ["size must be between 0 and 20", "must be a well-formed email address"])
...
servers:
  - url: http://localhost:8080
    description: Default endpoint for Kafbat UI API
    variables: {}